requirements.txt
asgiref==3.2.10
beautifulsoup4==4.9.3
boto3==1.14.20
botocore==1.17.20
certifi==2020.12.5
cffi==1.14.5
chardet==3.0.4
cryptography==3.4.7
dataclasses-json==0.5.2
defusedxml==0.7.1
dj-rest-auth==2.1.4
Django==3.1.4
django-allauth==0.44.0
django-rest-auth==0.9.5
django-storages==1.9.1
django-summernote==0.8.11.6
django-tables2==2.3.3
django-userforeignkey==0.3.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.6.0
docutils==0.15.2
gunicorn==20.0.4
idna==2.10
isodate==0.6.0
jmespath==0.10.0
jsonfield==3.1.0
marshmallow==3.10.0
marshmallow-enum==1.5.1
mypy-extensions==0.4.3
numpy==1.19.4
oauthlib==3.1.0
pandas==1.1.4
Pillow==8.2.0
psycopg2==2.8.6
pycparser==2.20
pydotplus==2.0.2
PyJWT==2.0.1
PyMySQL==1.0.2
pyparsing==2.4.7
python-dateutil==2.8.1
python-dotenv==0.15.0
python-youtube==0.7.0
python3-openid==3.2.0
pytube==10.6.0
pytz==2020.1
redis==3.5.3
requests==2.24.0
requests-oauthlib==1.3.0
s3transfer==0.3.3
sentry-sdk==0.19.5
six==1.15.0
soupsieve==2.2
sqlparse==0.3.1
stringcase==1.2.0
tablib==2.0.0
typing-extensions==3.7.4.3
typing-inspect==0.6.0
urllib3==1.25.9
xlrd==1.2.0
